Create quadratic A and B
This code takes inputs for values a, b, and c, enters them into the quadratic equation, and plots
the graph.

An extended version of this model could ask the user the value of a, b, and c with an input
prompt. Another improvement would be to base the values of the x-axis and y-axis on the values
in the parabola, so that it was never off the graph.